requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.2
python-jose[cryptography]==3.3.0
qrcode[pil]==7.4.2
